F01: Add implementation of ProjetService
